+++ test/tint/builtins/bufferArrayView/storage_sized_runtime_array.wgsl.expected.glsl
@@ -6,6 +6,6 @@ buffer v_block_1_ssbo {
 } v_1;
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
-  min(uint(0), (8u - 1u));
+  min(0u, (8u - 1u));
   v_1.inner[0u] = 2u;
 }
