+++ test/tint/builtins/bufferArrayView/uniform_struct.wgsl.expected.glsl
@@ -10,6 +10,6 @@ buffer out_block_1_ssbo {
 } v_2;
 layout(local_size_x = 1, local_size_y = 1, local_size_z = 1) in;
 void main() {
-  min(uint(0), (2u - 1u));
+  min(0u, (2u - 1u));
   v_2.inner = uintBitsToFloat(v_1.inner[2u].xy);
 }
